+++ source/wiz/package.py
@@ -13,6 +13,7 @@ from wiz import __version__
 import wiz.definition
 import wiz.mapping
 import wiz.symbol
+import wiz.history
 import wiz.exception
 
 
@@ -105,6 +106,11 @@ def extract_context(packages, environ_mapping=None):
 
     mapping = reduce(_combine, packages, dict(environ=environ_mapping or {}))
     mapping["environ"] = sanitise_environ_mapping(mapping.get("environ", {}))
+
+    wiz.history.record_action(
+        wiz.symbol.CONTEXT_EXTRACTION_ACTION,
+        packages=packages, initial=environ_mapping, context=mapping
+    )
     return mapping
 
 
@@ -378,7 +384,7 @@ class Package(wiz.mapping.Mapping):
             the elements from the *variant* will have priority.
 
         """
-        definition_data = definition.to_mapping()
+        definition_data = definition.to_dict()
         mapping = dict(
             (k, v) for k, v in definition_data.items() if k != "variants"
         )