+++ source/wiz/registry.py
@@ -16,7 +16,12 @@ def get_local():
 def get_defaults():
     """Return the default registries."""
     return [
-        os.path.join(os.sep, "mill3d", "server", "apps", "WIZ", "registry"),
+        os.path.join(
+            os.sep, "mill3d", "server", "apps", "WIZ", "registry", "primary"
+        ),
+        os.path.join(
+            os.sep, "mill3d", "server", "apps", "WIZ", "registry", "secondary"
+        ),
         os.path.join(os.sep, "jobs", "ads", ".wiz", "registry")
     ]
 
@@ -24,10 +29,10 @@ def get_defaults():
 def fetch(paths, include_local=True, include_working_directory=True):
     """Fetch all registries from *paths*.
 
-    *include_local* indicate whether the local registry should be included.
+    *include_local* indicates whether the local registry should be included.
 
-    *include_local* indicate whether the current working directory should be
-     parsed to discover a registry.
+    *include_working_directory* indicates whether the current working directory
+    should be parsed to discover a registry.
 
     """
     registries = []
@@ -46,21 +51,21 @@ def fetch(paths, include_local=True, include_working_directory=True):
     if registry_path and include_local:
         registries.append(registry_path)
 
-    registries.reverse()
     return registries
 
 
 def discover(path):
     """Return registry from *path* if available under the folder structure.
 
-    The registry path should be a ".registry" folder within *path*.
+    The registry path should be a :file:`.wiz/registry` folder within *path*.
 
-    Return the registry path discovered, or None if the register is not found
+    Return the registry path discovered, or None if the folder is not found
     or not accessible.
 
     .. note::
 
-        No registry will be fetched if *path* is not under `/jobs/ads`.
+        No registry folder will be fetched if *path* is not under
+        :file:`/jobs/ads`.
 
     """
     path = os.path.abspath(path)